% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_helper.R
\name{add_helper}
\alias{add_helper}
\title{Attaches a helper to a UI object.}
\usage{
add_helper(fun, content)
}
\arguments{
\item{fun}{\code{function} UI object to have helper attached}

\item{content}{\code{character} A title of a helper}
}
\value{
An UI object with helper attached
}
\description{
Attaches a helper to a UI object.
}
