% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_poland}
\alias{covid_poland}
\title{COVID-19 statistics in Poland.}
\format{
A data table with 132588 rows and 15 variables:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\source{
\url{https://www.gov.pl/web/koronawirus/wykaz-zarazen-koronawirusem-sars-cov-2}
}
\usage{
data(covid_poland)
}
\description{
A dataset includes daily measurements of COVID-19 virus spread in Poland.
The variables are as follows:
}
\keyword{datasets}
