% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_and_new_cases_hiv}
\alias{deaths_and_new_cases_hiv}
\title{Statistics on the global HIV / AIDS epidemic.}
\format{
A data table with 6468 rows and 6 variables:
\describe{
  \item{Entity}{name of the country / geographic area}
  \item{Code}{code of the geographical area}
  \item{Year}{year in which the measurements were made}
  \item{Deaths}{number of dead}
  \item{Incidence}{number of infected}
  \item{Prevalence}{number of infected per 100,000 people in the studied population}
}
}
\source{
\url{https://ourworldindata.org/hiv-aids}
}
\usage{
data(deaths_and_new_cases_hiv)
}
\description{
A dataset containing the number of cases and deaths caused by HIV around the world in 1990-2017
They include people of different ages and both genders. The variables are as follows:
}
\keyword{datasets}
