% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_column_check.R
\name{geo_column_check}
\alias{geo_column_check}
\title{Validating a given geographic column.}
\usage{
geo_column_check(vector_geo, data_type)
}
\arguments{
\item{vector_geo}{\code{character or numeric} selected column for validaling}

\item{data_type}{\code{character} type of geographical area - "Poland" or "World"}
}
\value{
logical value, TRUE when the data format is invalid
}
\description{
Validating a given geographic column.
}
