% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plotting leaflet map for selected dataset and columns.}
\usage{
plot_map(df, map_type, geo_column, date_column, measurements, data_value = NA)
}
\arguments{
\item{df}{\code{data.table} selected dataset for visualization}

\item{map_type}{\code{character} type of geographical area - "Poland" or "World"}

\item{geo_column}{\code{character} name of column contains geographical data}

\item{date_column}{\code{character} name of column contains time data}

\item{measurements}{\code{character} name of column contains measurements}

\item{data_value}{\code{character} a value of date_column for select rows to visualization}
}
\value{
map visualization
}
\description{
Plotting leaflet map for selected dataset and columns.
}
