% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poland_teryt}
\alias{poland_teryt}
\title{TERYT administrative codes.}
\format{
A data table with 395 rows and 3 variables:
\describe{
  \item{name}{name of the district or the city on district rights or the voivodeship}
  \item{administrative_division_level}{district or the city on district rights or the voivodeship}
  \item{teryt}{code of the geographical area}
}
}
\source{
\url{https://eteryt.stat.gov.pl/eTeryt/rejestr_teryt/udostepnianie_danych/baza_teryt/uzytkownicy_indywidualni/pobieranie/pliki_pelne.aspx?contrast=default}
}
\usage{
data(poland_teryt)
}
\description{
A dataset containing the TERYT Poland codes of the geographical area. The variables description:
}
\keyword{datasets}
