% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_tac_toe.R
\name{tic_tac_toe}
\alias{tic_tac_toe}
\title{tic_tac_toe}
\usage{
tic_tac_toe(User1, User2)
}
\arguments{
\item{User1}{character: nome del giocatore 1}

\item{User2}{character: nome del giocatore 2}
}
\value{
winner character: nome del vincitore
}
\description{
Avvia il gioco tris
}
\examples{
# Questo è un esempio
tic_tac_toe("Antonio","Carlotta")
}
\author{
Antonio Profico
}
