% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_column_check.R
\name{time_column_check}
\alias{time_column_check}
\title{Validating a given date/time column.}
\usage{
time_column_check(vector_time)
}
\arguments{
\item{vector_time}{\code{character or numeric} selected column for validaling}
}
\value{
logical value
}
\description{
Validating a given date/time column.
}
